feat: 예약 취소 시 예약 대기가 있으면 즉시 예약 기능 추가

+++ src/main/java/nextstep/reservation/service/ReservationServiceImpl.java
@@ -11,6 +11,8 @@ import nextstep.reservation.mapper.ReservationMapper;
 import nextstep.reservation.repository.ReservationRepository;
 import nextstep.schedule.dto.ScheduleResponse;
 import nextstep.schedule.service.ScheduleService;
+import nextstep.waiting.domain.Waiting;
+import nextstep.waiting.repository.WaitingRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,6 +26,8 @@ import java.util.stream.Collectors;
 public class ReservationServiceImpl implements ReservationService {
 
     private final ReservationRepository repository;
+
+    private final WaitingRepository waitingRepository;
 
     private final ScheduleService scheduleService;
 
@@ -68,8 +72,13 @@ public class ReservationServiceImpl implements ReservationService {
     public boolean deleteById(Long memberId, Long id) {
         Reservation reservation = repository.getById(id);
         validateReservation(reservation, memberId);
+        Waiting waiting = waitingRepository.getFirstByScheduleId(reservation.getScheduleId());
+        if (Objects.isNull(waiting)) {
+            return repository.deleteById(id);
+        }
+        waitingRepository.deleteById(waiting.getId());
 
-        return repository.deleteById(id);
+        return repository.updateById(id, waiting.getMemberId());
     }
 
     private void validateReservation(Reservation reservation, Long memberId) {
